+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parse `urn` scheme URIs.
 - Parse `file` scheme URIs.
 - Parse `http` scheme URIs.
+- Parse `https` scheme URIs.
 
 ## [1.0.3] - 2023-05-22